+++ baybe/parameters/__init__.py
@@ -1,6 +1,6 @@
 """BayBE parameters."""
 
-from baybe.parameters.base import ParameterMetadata
+from baybe.parameters.base import MeasurableMetadata
 from baybe.parameters.categorical import CategoricalParameter, TaskParameter
 from baybe.parameters.custom import CustomDiscreteParameter
 from baybe.parameters.enum import (
@@ -19,7 +19,7 @@ __all__ = [
     "CategoricalParameter",
     "CustomDiscreteParameter",
     "CustomEncoding",
-    "ParameterMetadata",
+    "MeasurableMetadata",
     "NumericalContinuousParameter",
     "NumericalDiscreteParameter",
     "SubstanceEncoding",